Added support for decimal timezones

+++ Jellyfin.Plugin.PlaybackReporting/Data/ActivityRepository.cs
@@ -393,7 +393,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             }
         }
 
-        public List<Dictionary<string, string>> GetUsageForUser(string date, string userId, string[] types, int timezoneOffset)
+        public List<Dictionary<string, string>> GetUsageForUser(string date, string userId, string[] types, float timezoneOffset)
         {
             List<string> filters = new List<string>();
             foreach (string filter in types)
@@ -441,7 +441,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             return items;
         }
 
-        public Dictionary<String, Dictionary<string, int>> GetUsageForDays(int days, DateTime endDate, string[] types, string? dataType, int timezoneOffset)
+        public Dictionary<String, Dictionary<string, int>> GetUsageForDays(int days, DateTime endDate, string[] types, string? dataType, float timezoneOffset)
         {
             List<string> filters = new List<string>();
             foreach (string filter in types)
@@ -505,7 +505,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             return usage;
         }
 
-        public SortedDictionary<string, int> GetHourlyUsageReport(int days, DateTime endDate, string[] types, int timezoneOffset)
+        public SortedDictionary<string, int> GetHourlyUsageReport(int days, DateTime endDate, string[] types, float timezoneOffset)
         {
             List<string> filters = new List<string>();
             foreach (string filter in types)
@@ -514,7 +514,6 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             }
 
             SortedDictionary<string, int> report_data = new SortedDictionary<string, int>();
-
             endDate = endDate.AddHours(-timezoneOffset);
             DateTime start_date = endDate.Subtract(new TimeSpan(days, 0, 0, 0));
 
@@ -573,7 +572,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             }
         }
 
-        public List<Dictionary<string, object>> GetBreakdownReport(int days, DateTime endDate, string type, int timezoneOffset)
+        public List<Dictionary<string, object>> GetBreakdownReport(int days, DateTime endDate, string type, float timezoneOffset)
         {
             // UserId ItemType PlaybackMethod ClientName DeviceName
 
@@ -658,7 +657,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             return report;
         }
 
-        public List<Dictionary<string, object>> GetTvShowReport(int days, DateTime endDate, int timezoneOffset)
+        public List<Dictionary<string, object>> GetTvShowReport(int days, DateTime endDate, float timezoneOffset)
         {
             List<Dictionary<string, object>> report = new List<Dictionary<string, object>>();
 
@@ -701,7 +700,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             return report;
         }
 
-        public List<Dictionary<string, object>> GetMoviesReport(int days, DateTime endDate, int timezoneOffset)
+        public List<Dictionary<string, object>> GetMoviesReport(int days, DateTime endDate, float timezoneOffset)
         {
             List<Dictionary<string, object>> report = new List<Dictionary<string, object>>();
 
@@ -742,7 +741,7 @@ namespace Jellyfin.Plugin.PlaybackReporting.Data
             return report;
         }
 
-        public List<Dictionary<string, object>> GetUserReport(int days, DateTime endDate, int timezoneOffset)
+        public List<Dictionary<string, object>> GetUserReport(int days, DateTime endDate, float timezoneOffset)
         {
             List<Dictionary<string, object>> report = new List<Dictionary<string, object>>();
